requirements.txt
asgiref==3.2.7
Django==3.0.4
django-sampledatahelper==0.5
djangorestframework==3.11.0
pytz==2019.3
sampledata==0.3.7
six==1.14.0
sqlparse==0.3.1
Unipath==1.1
